Add self.add_news_source to Article class and get called in Api self.make_articles method.
Article @news_source variable should be a Source object. When Api.make_articles gets called to create Article objects it does so with a news_source object as argument. This method takes article hashes and sends them to Article.new_from_json to instantiate objects but this does not include the Source object itself and so @news_source cannot be set with desired Source object.

By creating a Article.add_news_source method we can accomplish the above.  This is done so by calling this method during the Api.make_articles method and passing the news_source object there.